tensor_aio: Clarify statistics reset comment in memperfd
It's not immediately clear at a cursory glance why statistics are reset at
the very end of a memperfd run, considering that doing so requires taking
the PMU lock again.

Clarify the comment for the statistics reset in memperfd to indicate that
there's a good reason for why statistics aren't reset at the time that
memperfd reads them, and are instead reset at the very end of a run.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>