sched/fair: Enable scheduler feature NEXT_BUDDY
The NEXT_BUDDY feature reinforces wakeup preemption to encourage the last

wakee to be scheduled sooner on the assumption that the waker/wakee share

cache-hot data. In CFS, it was paired with LAST_BUDDY to switch back on

the assumption that the pair of tasks still share data but also relied

on START_DEBIT and the exact WAKEUP_PREEMPTION implementation to get

good results.

NEXT_BUDDY has been disabled since commit 0ec9fab3d186 ("sched: Improve

latencies and throughput") and LAST_BUDDY was removed in commit 5e963f2bd465

("sched/fair: Commit to EEVDF"). The reasoning is not clear but vruntime

spread is mentioned so a safe assumption is that NEXT_BUDDY had an impact

on overall fairness. It was not noted why LAST_BUDDY was removed but the

assumption is that it's very difficult to reason what LAST_BUDDY's correct

and effective behaviour should be while still respecting EEVDFs goals.

NEXT_BUDDY is relatively easier to reason about given that it's a

point-in-time decision on the wakees deadline and eligibilty relative to the

waker. Enable NEXT_BUDDY as a preparation path to document that the decision

to ignore the current implementation is deliberate. While not presented,

the results were at best neutral and often much more variable. It's known

that it has at least some users due to bug reports but less clear what

workloads benefit from the current implementation.

Signed-off-by: Mel Gorman <mgorman@techsingularity.net>